Move the implementation to a cpp file, to avoid douple definition errors (e.g. in NRF).
